KTB-15 added standalone class with telegram parser, use const as companion object
